Do CreateMigration and EvaluateDataLoss

+++ migration-engine/migration-engine-tests/src/commands/evaluate_data_loss.rs
@@ -58,7 +58,7 @@ impl std::fmt::Debug for EvaluateDataLossAssertion<'_> {
 
 impl<'a> EvaluateDataLossAssertion<'a> {
     #[track_caller]
-    pub fn assert_steps_count(self, count: usize) -> Self {
+    pub fn assert_steps_count(self, count: u32) -> Self {
         assert!(
             self.output.migration_steps == count,
             "Assertion failed. Expected evaluateDataLoss to return {} steps, found {}",
@@ -91,7 +91,7 @@ impl<'a> EvaluateDataLossAssertion<'a> {
         self
     }
 
-    pub fn assert_warnings_with_indices(self, warnings: &[(Cow<'_, str>, usize)]) -> Self {
+    pub fn assert_warnings_with_indices(self, warnings: &[(Cow<'_, str>, u32)]) -> Self {
         assert!(
             self.output.warnings.len() == warnings.len(),
             "Expected {} warnings, got {}.\n{:#?}",
@@ -100,7 +100,7 @@ impl<'a> EvaluateDataLossAssertion<'a> {
             self.output.warnings
         );
 
-        let descriptions: Vec<(Cow<'_, str>, usize)> = self
+        let descriptions: Vec<(Cow<'_, str>, u32)> = self
             .output
             .warnings
             .iter()
@@ -133,7 +133,7 @@ impl<'a> EvaluateDataLossAssertion<'a> {
         self
     }
 
-    pub fn assert_unexecutables_with_indices(self, unexecutables: &[(Cow<'_, str>, usize)]) -> Self {
+    pub fn assert_unexecutables_with_indices(self, unexecutables: &[(Cow<'_, str>, u32)]) -> Self {
         assert!(
             self.output.unexecutable_steps.len() == unexecutables.len(),
             "Expected {} unexecutables, got {}.\n{:#?}",
@@ -142,7 +142,7 @@ impl<'a> EvaluateDataLossAssertion<'a> {
             self.output.unexecutable_steps
         );
 
-        let descriptions: Vec<(Cow<'_, str>, usize)> = self
+        let descriptions: Vec<(Cow<'_, str>, u32)> = self
             .output
             .unexecutable_steps
             .iter()
